requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.10.4
pydantic-settings>=2.7.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Async and Task Queue
celery==5.3.4
redis==5.0.1
aioredis==2.0.1

# WebSockets
websockets==12.0
python-socketio==5.11.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
structlog==23.2.0
rich>=13.9.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI endpoints

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0

# Existing MCP dependencies
mcp-agent>=0.1.3
anthropic>=0.7.0
click>=8.0.0
pyyaml>=6.0
pathlib-mate>=1.0.0

# Node.js dependencies for MCP servers (install via npm)
# @modelcontextprotocol/server-memory
# @modelcontextprotocol/server-filesystem
# @modelcontextprotocol/server-fetch